requirements.txt
asgiref==3.4.1
backcall==0.2.0
bcrypt==3.2.0
cffi==1.14.6
decorator==5.0.9
Django==3.2.5
djangorestframework==3.12.4
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5